IDREF: Fourth implementation draft.

+++ sli/ingestion/ingestion-core/src/main/java/org/slc/sli/ingestion/handler/ReferenceResolutionHandler.java
@@ -11,6 +11,8 @@ import org.slf4j.LoggerFactory;
 
 import org.slc.sli.ingestion.ReferenceConstructor;
 import org.slc.sli.ingestion.ReferenceResolver;
+import org.slc.sli.ingestion.landingzone.IngestionFileEntry;
+import org.slc.sli.ingestion.util.MD5;
 import org.slc.sli.ingestion.validation.ErrorReport;
 
 /**
@@ -20,7 +22,7 @@ import org.slc.sli.ingestion.validation.ErrorReport;
  * @author tshewchuk
  *
  */
-public class ReferenceResolutionHandler extends AbstractIngestionHandler<File, File> {
+public class ReferenceResolutionHandler extends AbstractIngestionHandler<IngestionFileEntry, IngestionFileEntry> {
 
     private static final Logger LOG = LoggerFactory.getLogger(ReferenceResolutionHandler.class);
 
@@ -47,7 +49,7 @@ public class ReferenceResolutionHandler extends AbstractIngestionHandler<File, F
      * @param inputFilePath
      *            Full pathname of input XML file to be expanded.
      */
-    private void constructReferenceMap(Set<String> referenceSet, String inputFilePath) {
+    private boolean constructReferenceMap(Set<String> referenceSet, String inputFilePath) {
         long startTime = System.currentTimeMillis();
 
         ReferenceConstructor rc = new ReferenceConstructor(referenceSet);
@@ -55,9 +57,11 @@ public class ReferenceResolutionHandler extends AbstractIngestionHandler<File, F
             referenceObjects = rc.execute(inputFilePath);
         } catch (Exception e) {
             log("Cannot extract references from XML file " + xmlInputFile.getName());
+            return false;
         }
 
-        log("--- Reference Map Population took (ms): " + (System.currentTimeMillis() - startTime));
+        log(xmlInputFile.getName() + ": Reference Map Population took (ms): " + (System.currentTimeMillis() - startTime));
+        return true;
     }
 
     /**
@@ -71,20 +75,22 @@ public class ReferenceResolutionHandler extends AbstractIngestionHandler<File, F
      * @param outputFilePath
      *            Full pathname of expanded output XML file to be written.
      */
-    private void resolveReferences(Set<String> referenceSet, String inputFilePath, String outputFilePath) {
+    private boolean resolveReferences(Set<String> referenceSet, String inputFilePath, String outputFilePath) {
         long startTime = System.currentTimeMillis();
 
-        ReferenceResolver rr = new ReferenceResolver(referenceSet, referenceObjects);
         xmlOutputFile = new File(outputFilePath);
+        ReferenceResolver rr = new ReferenceResolver(referenceSet, referenceObjects);
         try {
             rr.execute(inputFilePath, outputFilePath);
         } catch (Exception e) {
             // Delete the output file and report the error.
             xmlOutputFile.delete();
             log("Cannot resolve references in XML file " + xmlInputFile.getName());
+            return false;
         }
 
-        log("--- XML Denormalization took (ms): " + (System.currentTimeMillis() - startTime));
+        log(xmlInputFile.getName() + ": XML Denormalization took (ms): " + (System.currentTimeMillis() - startTime));
+        return true;
     }
 
     private void log(String errorMessage) {
@@ -103,22 +109,35 @@ public class ReferenceResolutionHandler extends AbstractIngestionHandler<File, F
      *            Error report for reporting errors/warnings.
      */
     @Override
-    File doHandling(File item, ErrorReport errorReport) {
+    IngestionFileEntry doHandling(IngestionFileEntry item, ErrorReport errorReport) {
         // Create the expanded output file.
-        xmlInputFile = item;
+        xmlInputFile = item.getFile();
         this.errorReport = errorReport;
         String inputFilePathname = xmlInputFile.getPath();
         String outputFilePathname = inputFilePathname.substring(0, inputFilePathname.lastIndexOf(".xml"))
                 + "_RESOLVED.xml";
-        constructReferenceMap(referenceObjectList, inputFilePathname);
-        resolveReferences(referenceObjectList, inputFilePathname, outputFilePathname);
 
-        // Move the expanded output file to the input file, and return it.
-        xmlInputFile.delete();
-        if (!xmlOutputFile.renameTo(xmlInputFile)) {
-            log("Error renaming " + xmlOutputFile.getName() + " to " + xmlInputFile.getName());
+        // First, construct the reference list.
+        boolean referenceMapConstructed = constructReferenceMap(referenceObjectList, inputFilePathname);
+        if (referenceMapConstructed) {
+            // Next, resolve the references.
+            boolean referencesResolved = resolveReferences(referenceObjectList, inputFilePathname, outputFilePathname);
+            if (referencesResolved) {
+                // Move the expanded output file to the input file, and return it.
+                xmlInputFile.delete();
+                if (!xmlOutputFile.renameTo(xmlInputFile)) {
+                    log("Error renaming " + xmlOutputFile.getName() + " to " + xmlInputFile.getName());
+                }
+                IngestionFileEntry outputFileEntry = new IngestionFileEntry(item.getFileFormat(), item.getFileType(),
+                        xmlOutputFile.getName(), MD5.calculate(xmlOutputFile));
+                outputFileEntry.setFile(xmlOutputFile);
+                return outputFileEntry;
+            } else {
+                return null;
+            }
+        } else {
+            return null;
         }
-        return xmlOutputFile;
     }
 
     public void setReferenceObjectList(Set<String> referenceObjectList) {
